requirements.txt
numpy==1.26.4
torch===1.13.1
torchvision==0.14.1
transformers==4.37.2
pillow==11.0.0
pycocotools==2.0.8
accelerate==0.21.0
peft==0.6.1
sentencepiece==0.2.0
regex==2024.11.6
omegaconf==2.3.0
pandas==2.2.3
iopath==0.1.10
timm==0.9.2
opencv-python==4.10.0.84
webdataset==0.2.100
scikit-image==0.25.0
visual-genome==1.1.1
decord==0.6.0
wandb==0.19.1
ftfy==6.3.1
git+https://github.com/openai/CLIP.git
einops==0.8.0